\name{nSGMM-package}
\alias{nSGMM-package}
\alias{nSGMM}
\docType{package}
\title{
\packageTitle{nSGMM}
}
\description{
\packageDescription{nSGMM}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{nSGMM}
\packageIndices{nSGMM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{nSGMM}

Maintainer: \packageMaintainer{nSGMM}
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
